Import top GWAS/QTL summary statistics

The resulting top_SNPs data.frame can be used to guide the finemap_loci in querying and fine-mapping loci.

import_topSNPs(
  topSS,
  show_table = TRUE,
  sheet = 1,
  munge = TRUE,
  chrom_col = NULL,
  position_col = NULL,
  min_POS_col = NULL,
  max_POS_col = NULL,
  snp_col = NULL,
  pval_col = NULL,
  effect_col = NULL,
  locus_col = "Locus",
  gene_col = "Gene",
  grouping_vars = c("Locus"),
  remove_variants = NULL,
  verbose = TRUE
)

Arguments

topSS

Can be a data.frame with the top summary stats per locus. Alternatively, you can provide a path to the stored top summary stats file. Can be in any tabular format (e.g. excel, .tsv, .csv, etc.). This file should have one lead GWAS/QTL hits per locus. If there is more than one SNP per locus, the one with the smallest p-value (then the largest effect size) is selected as the lead SNP. The lead SNP will be used as the center of the locus when constructing the locus subset files.

show_table

Create an interative data table.

sheet

If the topSS file is an excel sheet, you can specify which tab to use. You can provide either a number to identify the tab by order, or a string to identify the tab by name.

munge

Standardise column names.

chrom_col

Column containing chromosome (can be "chr1" or "1" format).

position_col

Column containing genomic position in units of basepairs.

min_POS_col

Column containing minimum genomic position (used instead of an arbitrary window size).

max_POS_col

Column containing maximum genomic position (used instead of an arbitrary window size).

snp_col

Column containing SNP RSIDs.

pval_col

Column containing uncorrected p-values.

effect_col

Column containing effect size (e.g. Beta, OR).

locus_col

Column containing unique locus name.

gene_col

An optional column to keep track of the causal gene(s) in each locus (if known).

grouping_vars

The variables that you want to group by such that each grouping_var combination has its own index SNP. For example, if you want one index SNP per QTL eGene - GWAS locus pair, you could supply: grouping_vars=c("Locus","Gene").

remove_variants

SNPs to remove from topSS,

verbose

Print messages.

Value

Munged topSNPs table.

Examples

topSNPs <- echodata::import_topSNPs(
    topSS = echodata::topSNPs_Nalls2019_raw, 
    pval_col = "P, all studies",
    effect_col = "Beta, all studies",
    locus_col = "Nearest Gene", 
    gene_col = "QTL Nominated Gene (nearest QTL)")
#> Renaming column: P, all studies ==> P
#> Renaming column: Beta, all studies ==> Effect
#> Renaming column: Nearest Gene ==> Locus
#> Renaming column: QTL Nominated Gene (nearest QTL) ==> Gene
#> [1] "+ Assigning gene_col and locus_col independently."
#> Standardising column headers.
#> First line of summary statistics file: 
#> SNP	CHR	BP	Locus	Gene	Effect allele	Other allele	Effect allele frequency	Effect	SE, all studies	P	P, COJO, all studies	P, random effects, all studies	P, Conditional 23AndMe only	P, 23AndMe only	I2, all studies	Freq1, previous studies	Beta, previous studies	StdErr, previous studies	P, previous studies	I2, previous studies	Freq1, new studies	Beta, new studies	StdErr, new studies	P, new studies	I2, new studies	Passes pooled 23andMe QC	Known GWAS locus within 1MB	Failed final filtering and QC	Locus within 250KB	Locus Number	
#> Returning unmapped column names without making them uppercase.
#> + Mapping colnames from MungeSumstats ==> echolocatoR