Search and download fine-mapping files

Search the echolocatoR Fine-mapping Portal for fine-mapping results, LD, and locus plots.

portal_query(
  dataset_types = NULL,
  datasets = NULL,
  phenotypes = NULL,
  loci = NULL,
  LD_panels = c("UKB", "1KGphase1", "1KGphase3"),
  file_types = c("multi_finemap", "LD", "plot"),
  results_dir = tempdir(),
  overwrite = FALSE,
  nThread = 1,
  verbose = TRUE
)

Arguments

dataset_types
Dataset types to search for (e.g. "GWAS", "QTL").

datasets
Dataset names to search for (e.g. "Ripke_2014", "Wray_2018").

phenotypes
Phenotypes to search for (e.g. "Schizophrenia", "Major Depressive Disorder").

loci
Locus names to search for (e.g. "BST1", "CHRNB1", "LRRK2").

LD_panels
LD panels to search for, or results generated using different LD panels (e.g. "UKB", "1KGphase1", "1KGphase3").

file_types
File types to search for:

"multi_finemap":Fine-mapping results merged with GWAS/QTL summary statistics.

"LD":Linkage Disequilibrium with the lead SNP in each locus.

"plot" : Locus plots of fine-mapped results.

results_dir
Where to save the matching queries.

overwrite
Whether to overwrite previously saved queries with the same names.

nThread
Number of threads to parallelise downloads across.

verbose
Print messages.

Value

List of local paths where the requested files were downloaded to.

Examples

if (FALSE) {
local_finemap <- portal_query(
    dataset_types = "GWAS",
    phenotypes = c("schizophrenia", "parkinson"),
    file_types = "multi_finemap",
    loci = c("BST1", "CHRNB1", "LRRK2"),
    LD_panels = "1KGphase3"
)
}